refactor: organize demo view parts
