+++ demo/src/app/view/demo-view/parts/demo-parts-center/demo-parts-center.ts
