+++ demo/src/app/view/demo-view/parts/demo-parts-log/demo-parts-log.html
