+++ demo/src/app/view/demo-view/parts/demo-parts-log/demo-parts-log.ts
@@ -1,5 +1,5 @@
 import { Component, inject } from '@angular/core';
-import { DemoState } from '../../../domain/state/global/demo-global.state';
+import { DemoState } from '../../../../domain/state/global/demo-global.state';
 
 @Component({
   selector: 'app-demo-parts-log',
