+++ demo/src/app/view/demo-view/parts/demo-parts-select/demo-parts-select.scss
